requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0

# Redis (optional)
redis[hiredis]==5.0.1

# AI/ML libraries - Updated versions to resolve conflicts
torch>=2.3.0,<3.0
torchvision>=0.18.0
torchaudio>=2.3.0
stable-baselines3==2.7.0
scikit-learn==1.3.2
pandas==2.1.3
numpy>=1.24.4,<2.0
ta==0.10.2

# Cryptocurrency libraries - Updated aiohttp for ccxt compatibility
ccxt>=4.4.99
python-binance==1.0.19

# Async libraries - Updated aiohttp version
aiohttp>=3.10.11
asyncio-throttle==1.0.2

# Utilities
loguru==0.7.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional dependencies for proper type hints and imports
typing-extensions>=4.0.0
